+++ sdk/eventhub/arm-eventhub/samples/v6-beta/javascript/applicationGroupOperationsDeleteSample.js
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { EventHubManagementClient } from "@azure/arm-eventhub";
-import { DefaultAzureCredential } from "@azure/identity";
+const { EventHubManagementClient } = require("@azure/arm-eventhub");
+const { DefaultAzureCredential } = require("@azure/identity");
 
 /**
  * This sample demonstrates how to deletes an ApplicationGroup for a Namespace.
@@ -10,14 +10,18 @@ import { DefaultAzureCredential } from "@azure/identity";
  * @summary deletes an ApplicationGroup for a Namespace.
  * x-ms-original-file: 2026-01-01/ApplicationGroup/ApplicationGroupDelete.json
  */
-async function applicationGroupDelete(): Promise<void> {
+async function applicationGroupDelete() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new EventHubManagementClient(credential, subscriptionId);
-  await client.applicationGroup.delete("contosotest", "contoso-ua-test-eh-system-1", "appGroup1");
+  await client.applicationGroupOperations.delete(
+    "contosotest",
+    "contoso-ua-test-eh-system-1",
+    "appGroup1",
+  );
 }
 
-async function main(): Promise<void> {
+async function main() {
   await applicationGroupDelete();
 }