+++ sdk/eventhub/arm-eventhub/samples/v6-beta/javascript/applicationGroupOperationsGetSample.js
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { EventHubManagementClient } from "@azure/arm-eventhub";
-import { DefaultAzureCredential } from "@azure/identity";
+const { EventHubManagementClient } = require("@azure/arm-eventhub");
+const { DefaultAzureCredential } = require("@azure/identity");
 
 /**
  * This sample demonstrates how to gets an ApplicationGroup for a Namespace.
@@ -10,11 +10,11 @@ import { DefaultAzureCredential } from "@azure/identity";
  * @summary gets an ApplicationGroup for a Namespace.
  * x-ms-original-file: 2026-01-01/ApplicationGroup/ApplicationGroupGet.json
  */
-async function applicationGroupGet(): Promise<void> {
+async function applicationGroupGet() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new EventHubManagementClient(credential, subscriptionId);
-  const result = await client.applicationGroup.get(
+  const result = await client.applicationGroupOperations.get(
     "contosotest",
     "contoso-ua-test-eh-system-1",
     "appGroup1",
@@ -22,7 +22,7 @@ async function applicationGroupGet(): Promise<void> {
   console.log(result);
 }
 
-async function main(): Promise<void> {
+async function main() {
   await applicationGroupGet();
 }