+++ sdk/eventhub/arm-eventhub/samples/v6-beta/typescript/src/applicationGroupOperationsDeleteSample.ts
@@ -5,28 +5,24 @@ import { EventHubManagementClient } from "@azure/arm-eventhub";
 import { DefaultAzureCredential } from "@azure/identity";
 
 /**
- * This sample demonstrates how to gets a list of application groups for a Namespace.
+ * This sample demonstrates how to deletes an ApplicationGroup for a Namespace.
  *
- * @summary gets a list of application groups for a Namespace.
- * x-ms-original-file: 2026-01-01/ApplicationGroup/ApplicationGroupListByNamespace.json
+ * @summary deletes an ApplicationGroup for a Namespace.
+ * x-ms-original-file: 2026-01-01/ApplicationGroup/ApplicationGroupDelete.json
  */
-async function listApplicationGroups(): Promise<void> {
+async function applicationGroupDelete(): Promise<void> {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new EventHubManagementClient(credential, subscriptionId);
-  const resArray = new Array();
-  for await (const item of client.applicationGroup.listByNamespace(
+  await client.applicationGroupOperations.delete(
     "contosotest",
     "contoso-ua-test-eh-system-1",
-  )) {
-    resArray.push(item);
-  }
-
-  console.log(resArray);
+    "appGroup1",
+  );
 }
 
 async function main(): Promise<void> {
-  await listApplicationGroups();
+  await applicationGroupDelete();
 }
 
 main().catch(console.error);