+++ sdk/eventhub/arm-eventhub/samples/v6-beta/typescript/src/applicationGroupOperationsGetSample.ts
@@ -14,7 +14,7 @@ async function applicationGroupGet(): Promise<void> {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new EventHubManagementClient(credential, subscriptionId);
-  const result = await client.applicationGroup.get(
+  const result = await client.applicationGroupOperations.get(
     "contosotest",
     "contoso-ua-test-eh-system-1",
     "appGroup1",