+++ sdk/eventhub/arm-eventhub/src/api/networkSecurityPerimeterConfigurationOperations/operations.ts
@@ -8,7 +8,7 @@ import {
   networkSecurityPerimeterConfigurationListDeserializer,
 } from "../../models/models.js";
 import { expandUrlTemplate } from "../../static-helpers/urlTemplate.js";
-import type { NetworkSecurityPerimeterConfigurationListOptionalParams } from "./options.js";
+import type { NetworkSecurityPerimeterConfigurationOperationsListOptionalParams } from "./options.js";
 import type { StreamableMethod, PathUncheckedResponse } from "@azure-rest/core-client";
 import { createRestError, operationOptionsToRequestParameters } from "@azure-rest/core-client";
 
@@ -16,7 +16,9 @@ export function _listSend(
   context: Client,
   resourceGroupName: string,
   namespaceName: string,
-  options: NetworkSecurityPerimeterConfigurationListOptionalParams = { requestOptions: {} },
+  options: NetworkSecurityPerimeterConfigurationOperationsListOptionalParams = {
+    requestOptions: {},
+  },
 ): StreamableMethod {
   const path = expandUrlTemplate(
     "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.EventHub/namespaces/{namespaceName}/networkSecurityPerimeterConfigurations{?api%2Dversion}",
@@ -57,7 +59,9 @@ export async function list(
   context: Client,
   resourceGroupName: string,
   namespaceName: string,
-  options: NetworkSecurityPerimeterConfigurationListOptionalParams = { requestOptions: {} },
+  options: NetworkSecurityPerimeterConfigurationOperationsListOptionalParams = {
+    requestOptions: {},
+  },
 ): Promise<NetworkSecurityPerimeterConfigurationList> {
   const result = await _listSend(context, resourceGroupName, namespaceName, options);
   return _listDeserialize(result);